operating_platform/robot/components/arm_normal_uarm_v1/pyproject.toml
[project]
name = "dr-component-arm-uarm"
version = "0.1.0"
description = "Anything-U-Arm Leader Arm Component"
authors = [
    {name = "Xiang Yang",email = "ryu-yang@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "dora-rs-cli (>=0.3.11,<0.4.0)",
    "feetech-servo-sdk>=1.0.0",
    "numpy",
    "deepdiff",
    "tqdm",
    "pygame",
    "draccus",
    "pynput"
]

[tool.setuptools]
packages = []

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

